applied http://savannah.nongnu.org/patch/?func=detailitem&item_id=4814 (Daniel Hokka Zakrisson):
This patch lets users configure disk limits in
/etc/vservers/<name>/dlimits/*/, and have them set on the start of the
vserver, as well as saved when stopping the vserver.

For counting the amount of inodes and space used by a certain vserver
on a filesystem, I modified the old and broken vdu program (vdu bin
etc would not find etc, for instance) and made it useful again.

The cache files will probably need some love, I'm a bit uncertain as
to where to put them (or rather, the symlink to their location).

This is the last piece of the disk limits puzzle. A variant of this
would be really nice to have in 0.30.210. (do you have an ETA on that,
btw?)


git-svn-id: http://svn.linux-vserver.org/svn/util-vserver/trunk@2252 94cd875c-1c1d-0410-91d2-eb244daf1a30