vserver: trigger automount on 'vserver start'
Call 'secure-mount --trigger-automount' before entering the new
namespace.  This is useful in combination with automounters which
operate in host namespace and can not send the new mountpoints into
the guest context.

E.g. assuming a systemd system with

| /dev/sda1  /mnt/foo   ext4    x-systemd.automount 1 1

in its /etc/fstab and a vserver which tries to bind mount

| /mnt/foo   /mnt       bind

Then on system boot:

1. /mnt/foo is not mounted yet

2. 'vserver foo start' creates a new namespace and tries to bind mount
   /mnt/foo

3. systemd sees the automount trigger and mounts /dev/sda1

4. new mount is visible in host fs but not in the vserver

Patch adds an extra step between 1. and 2. where '/mnt/foo/.' is
accessed so that /dev/sda1 is mounted before the new namespace is
created.

Signed-off-by: Enrico Scholz <enrico.scholz@ensc.de>