secure-mount: fix the file-mount support
Decision whether file-mount is required was done by looking at
destination object while being the rootfs.  This is unexpected and
will fail because rootfs usually does not have the fileystem layout
of the guest.

To fix this, check the type of the source.  Alternatively, the check
can be delayed into the mountSingle() function but this change is more
invasive...

Signed-off-by: Enrico Scholz <enrico.scholz@ensc.de>