fixup! secure-mount: fix the file-mount support
Execute the is-file check only for bind mounts; they might trigger
e.g. for loop mounts too which is unwanted. As src is located in host
filesystem, we can allow symlinks there.

To prevent symlink attacks (which were possible with the previous
dst-check too) we have to verify the mount destination.  Although the
patch allows races when executed for running guests, it is safe in the
'vserver start' case.


Signed-off-by: Enrico Scholz <enrico.scholz@ensc.de>